rename: 예시용 코드 접두사로 example 추가
